cambio enrutamiento para incluir el rol de tecnico
