+++ src/main/java/acme/features/technician/tasks/TechnicianTaskGuiController.java
@@ -1,4 +1,4 @@
-package acme.features.tasks;
+package acme.features.technician.tasks;
 
 import javax.annotation.PostConstruct;
